requirements.txt
django
python-dotenv
psycopg2-binary
django-phonenumber-field[phonenumberslite]
djangorestframework
markdown
django-filter
pillow